+++ tests/sync/test_organization.py
@@ -4,7 +4,7 @@ from pydantic import ValidationError
 
 from checkbox_sdk.client.synchronous import CheckBoxClient
 from checkbox_sdk.exceptions import CheckBoxAPIError
-from ..models.organization_models import ReceiptConfigShema, SmsBillingSchema
+from ..models.organization_models import ReceiptConfigShema, SmsBillingSchema, BillingStatusSchema
 
 
 def test_organization_receipt_config(auth_token, license_key):
@@ -55,3 +55,18 @@ def test_organization_sms_billing(auth_token, license_key):
             assert model is not None
         except ValidationError as e:  # pragma: no cover
             pytest.fail(f"Sms billing validation schema failed: {e}")
+
+
+def test_organization_billing_status(auth_token, license_key):
+    with CheckBoxClient() as client:
+        client.cashier.authenticate_token(auth_token, license_key=license_key)
+
+        assert client.storage.cash_register["is_test"], "Not test cash register"
+
+        # sourcery skip: no-loop-in-tests
+        sms_billing = client.organization.get_organization_billing_status()
+        try:
+            model = BillingStatusSchema(**sms_billing)
+            assert model is not None
+        except ValidationError as e:  # pragma: no cover
+            pytest.fail(f"Billing status validation schema failed: {e}")