+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [1.3.0] - 2025-06-09
+
+### Added
+
+- Support for new GET /api/v1/organization/billing-status method.
+  This endpoint allows clients to retrieve the current billing status of an organization.
+  [More details](https://checkbox.ua/blog/novi-funktsii-checkbox-u-travni-2025/)
+
 ## [1.2.0] - 2025-02-13
 
 ### Added